Outsource localStorage management
By extracting the feature as a reusable hook, you can use it in more than one React component. Every component just needs to use a unique storageKey to not collide with other component's storages.
Anyway, even though this custom hook shows you the idea of how it works, you should rely on an open source variant of it for your React production application.
https://www.npmjs.com/package/use-local-storage-state